+++ lib/assets/javascripts/new-dashboard/components/MapsList.vue
@@ -9,16 +9,16 @@
         <template slot="title">
           <VisualizationsTitle
             :defaultTitle="$t(`MapsPage.header.title['${appliedFilter}']`)"
-            :selectedItems="selectedMaps.length"
-            :vizQuota="publicMapsQuota"
-            :vizCount="publicMapsCount"
-            :isOutOfQuota="isOutOfPublicMapsQuota"
-            :counterLabel="'Public Maps'"/>
+            :selectedItems="selectedMaps.length"/>
         </template>
 
-        <template v-if="shouldShowLimitsWarning" slot="warning">
-          <NotificationBadge type="warning">
-            <div class="warning" v-html="$t('MapsPage.header.warning', { counter: `${publicMapsCount}/${publicMapsQuota}`, path: upgradeUrl })"></div>
+        <template slot="warning">
+          <NotificationBadge type="warning" v-if="shouldShowLimitsWarning">
+            <div class="warning">
+              <span v-if="isOutOfPublicMapsQuota" class="is-bold" v-html="$t('MapsPage.header.warning.counter', { counter: `${publicMapsCount}/${publicMapsQuota}`, type: `public` })"></span>
+              <span v-if="isOutOfPrivateMapsQuota" class="is-bold" v-html="$t('MapsPage.header.warning.counter', { counter: `${privateMapsCount}/${privateMapsQuota}`, type: `private` })"></span>
+              <span v-html="$t('MapsPage.header.warning.upgrade', { path: upgradeUrl })"></span>
+            </div>
           </NotificationBadge>
         </template>
 
@@ -178,7 +178,10 @@ export default {
     ...mapGetters({
       publicMapsQuota: 'user/publicMapsQuota',
       publicMapsCount: 'user/publicMapsCount',
-      isOutOfPublicMapsQuota: 'user/isOutOfPublicMapsQuota'
+      isOutOfPublicMapsQuota: 'user/isOutOfPublicMapsQuota',
+      privateMapsQuota: 'user/privateMapsQuota',
+      privateMapsCount: 'user/privateMapsCount',
+      isOutOfPrivateMapsQuota: 'user/isOutOfPrivateMapsQuota'
     }),
     areAllMapsSelected () {
       return Object.keys(this.maps).length === this.selectedMaps.length;
@@ -213,7 +216,7 @@ export default {
       return this.selectedMaps.length > 0;
     },
     shouldShowLimitsWarning () {
-      return !this.selectedMaps.length && this.isOutOfPublicMapsQuota;
+      return this.isOutOfPublicMapsQuota || this.isOutOfPrivateMapsQuota;
     },
     shouldShowViewSwitcher () {
       return this.canChangeViewMode && !this.initialState && !this.emptyState && !this.selectedMaps.length;