+++ NEWS.md
@@ -25,6 +25,9 @@ Development
 - New Free Plan (Dashboard): Update metrics section in homepage ([#15435](https://github.com/CartoDB/cartodb/issues/15435))
 - New Free Plan (Dashboard): Update welcome section ([#15434](https://github.com/CartoDB/cartodb/issues/15434))
 - New Free Plan (Dashboard): Update custom API keys section in settings ([#15440](https://github.com/CartoDB/cartodb/issues/15440))
+- New Free Plan (Dashboard): Update change privacy and duplicate options ([#15441](https://github.com/CartoDB/cartodb/issues/15441))
+- New Free Plan (Dashboard): Update limits counter and warnings in maps section ([#15436](https://github.com/CartoDB/cartodb/issues/15436))
+- New Free Plan (Dashboard): Update limits counter and warnings in datasets section ([#15439](https://github.com/CartoDB/cartodb/issues/15439))
 
 4.34.0 (2020-01-28)
 -------------------